NP-244: Moving constants into their own page, and cleaning up imports
