+++ app/uk/gov/hmrc/uknwauthcheckerapi/models/constants/ApiErrorMessages.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.uknwauthcheckerapi.models
+package uk.gov.hmrc.uknwauthcheckerapi.models.constants
 
 object ApiErrorMessages {
   val badRequest:               String = "Bad request"
