+++ app/uk/gov/hmrc/uknwauthcheckerapi/models/constants/CustomRegexes.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.uknwauthcheckerapi.models
+package uk.gov.hmrc.uknwauthcheckerapi.models.constants
 
 import scala.util.matching.Regex
