+++ app/uk/gov/hmrc/uknwauthcheckerapi/models/constants/HmrcMimeTypes.scala
@@ -14,16 +14,8 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.uknwauthcheckerapi.models
+package uk.gov.hmrc.uknwauthcheckerapi.models.constants
 
-object ApiErrorCodes {
-  val badRequest: String = "BAD_REQUEST"
-  val forbidden:  String = "FORBIDDEN"
-  val internalServerError = "INTERNAL_SERVER_ERROR"
-  val invalidFormat:            String = "INVALID_FORMAT"
-  val matchingResourceNotFound: String = "MATCHING_RESOURCE_NOT_FOUND"
-  val methodNotAllowed:         String = "METHOD_NOT_ALLOWED"
-  val notAcceptable:            String = "NOT_ACCEPTABLE"
-  val serviceUnavailable:       String = "SERVICE_UNAVAILABLE"
-  val unauthorized:             String = "UNAUTHORIZED"
+object HmrcMimeTypes {
+  val json = "application/vnd.hmrc.1.0+json"
 }
