+++ app/uk/gov/hmrc/uknwauthcheckerapi/services/IntegrationFrameworkService.scala
@@ -29,7 +29,7 @@ import uk.gov.hmrc.uknwauthcheckerapi.config.AppConfig
 import uk.gov.hmrc.uknwauthcheckerapi.connectors.IntegrationFrameworkConnector
 import uk.gov.hmrc.uknwauthcheckerapi.errors.DataRetrievalError
 import uk.gov.hmrc.uknwauthcheckerapi.errors.DataRetrievalError._
-import uk.gov.hmrc.uknwauthcheckerapi.models.CustomRegexes._
+import uk.gov.hmrc.uknwauthcheckerapi.models.constants.CustomRegexes._
 import uk.gov.hmrc.uknwauthcheckerapi.models.eis.{EisAuthorisationRequest, EisAuthorisationResponseError}
 import uk.gov.hmrc.uknwauthcheckerapi.models.{AuthorisationRequest, AuthorisationResponse, AuthorisationsResponse}
