+++ it/test/uk/gov/hmrc/uknwauthcheckerapi/controllers/AuthorisationControllerISpec.scala
@@ -25,10 +25,10 @@ import play.api.libs.json.Json
 import play.api.libs.ws.WSResponse
 import play.mvc.Http.HeaderNames
 import uk.gov.hmrc.uknwauthcheckerapi.BaseISpec
-import uk.gov.hmrc.uknwauthcheckerapi.generators.{InvalidEorisAuthorisationRequest, TestRegexes, UtcDateTime, ValidAuthorisationRequest}
-import uk.gov.hmrc.uknwauthcheckerapi.models.eis.EisAuthorisationsResponse
-import uk.gov.hmrc.uknwauthcheckerapi.models.{AuthorisationRequest, CustomHeaderNames}
-import uk.gov.hmrc.uknwauthcheckerapi.utils.{EisAuthTypes, HmrcContentTypes}
+import uk.gov.hmrc.uknwauthcheckerapi.generators._
+import uk.gov.hmrc.uknwauthcheckerapi.models._
+import uk.gov.hmrc.uknwauthcheckerapi.models.constants.{CustomHeaderNames, HmrcContentTypes}
+import uk.gov.hmrc.uknwauthcheckerapi.models.eis._
 
 class AuthorisationControllerISpec extends BaseISpec {
