+++ test/uk/gov/hmrc/uknwauthcheckerapi/controllers/AuthActionSpec.scala
@@ -26,7 +26,7 @@ import uk.gov.hmrc.auth.core.retrieve.Credentials
 import uk.gov.hmrc.auth.core.{InternalError, MissingBearerToken}
 import uk.gov.hmrc.uknwauthcheckerapi.controllers.actions.AuthAction
 import uk.gov.hmrc.uknwauthcheckerapi.errors.{ServiceUnavailableApiError, UnauthorizedApiError}
-import uk.gov.hmrc.uknwauthcheckerapi.models.ApiErrorMessages
+import uk.gov.hmrc.uknwauthcheckerapi.models.constants.ApiErrorMessages
 
 class AuthActionSpec extends BaseSpec {
