+++ test/uk/gov/hmrc/uknwauthcheckerapi/controllers/AuthorisationsControllerSpec.scala
@@ -31,10 +31,10 @@ import play.api.test.Helpers._
 import uk.gov.hmrc.auth.core.AuthConnector
 import uk.gov.hmrc.uknwauthcheckerapi.errors.DataRetrievalError._
 import uk.gov.hmrc.uknwauthcheckerapi.errors._
-import uk.gov.hmrc.uknwauthcheckerapi.generators.{NoEorisAuthorisationRequest, TooManyEorisAuthorisationRequest, UtcDateTime, ValidAuthorisationRequest}
-import uk.gov.hmrc.uknwauthcheckerapi.models.{AuthorisationRequest, AuthorisationResponse, AuthorisationsResponse}
-import uk.gov.hmrc.uknwauthcheckerapi.services.{IntegrationFrameworkService, LocalDateService, ValidationService}
-import uk.gov.hmrc.uknwauthcheckerapi.utils.{ErrorMessages, JsonErrors}
+import uk.gov.hmrc.uknwauthcheckerapi.generators._
+import uk.gov.hmrc.uknwauthcheckerapi.models._
+import uk.gov.hmrc.uknwauthcheckerapi.models.constants._
+import uk.gov.hmrc.uknwauthcheckerapi.services._
 
 class AuthorisationsControllerSpec extends BaseSpec {
 
@@ -88,7 +88,7 @@ class AuthorisationsControllerSpec extends BaseSpec {
 
       val jsError = JsError(
         Seq("date", "eoris").map { field =>
-          (JsPath \ field, Seq(JsonValidationError(JsonErrors.pathMissing)))
+          (JsPath \ field, Seq(JsonValidationError(JsonErrorMessages.pathMissing)))
         }
       )
 
@@ -110,7 +110,7 @@ class AuthorisationsControllerSpec extends BaseSpec {
       val request = fakeRequestWithJsonBody(emptyJson)
 
       val jsError = JsError(
-        Seq((JsPath \ "", Seq(JsonValidationError(JsonErrors.expectedJsObject))))
+        Seq((JsPath \ "", Seq(JsonValidationError(JsonErrorMessages.expectedJsObject))))
       )
 
       val expectedResponse = Json.toJson(
@@ -200,7 +200,7 @@ class AuthorisationsControllerSpec extends BaseSpec {
   "return BAD_REQUEST (400) with authorised eoris when request has valid eoris but they exceed the maximum eoris" in {
 
     forAll { authorisationRequest: TooManyEorisAuthorisationRequest =>
-      val jsError = JsError(JsPath \ "eoris", JsonValidationError(ErrorMessages.invalidEoriCount))
+      val jsError = JsError(JsPath \ "eoris", JsonValidationError(ApiErrorMessages.invalidEoriCount))
 
       val expectedResponse = Json.toJson(
         JsonValidationApiError(jsError)
@@ -222,7 +222,7 @@ class AuthorisationsControllerSpec extends BaseSpec {
   "return BAD_REQUEST (400) when request has no eoris" in {
 
     forAll { authorisationRequest: NoEorisAuthorisationRequest =>
-      val jsError = JsError(JsPath \ "eoris", JsonValidationError(ErrorMessages.invalidEoriCount))
+      val jsError = JsError(JsPath \ "eoris", JsonValidationError(ApiErrorMessages.invalidEoriCount))
 
       val expectedResponse = Json.toJson(
         JsonValidationApiError(jsError)
